requirements.txt
google-cloud-aiplatform
streamlit
pillow